+++ src/mm_machine_mgr.h
@@ -14,7 +14,7 @@ struct mm_machinemgr
 	pthread_spinlock_t lock;
 	mm_list_t          list;
 	int                count;
-	int                seq;
+	uint64_t           seq;
 };
 
 void mm_machinemgr_init(mm_machinemgr_t*);
@@ -22,6 +22,6 @@ void mm_machinemgr_free(mm_machinemgr_t*);
 void mm_machinemgr_add(mm_machinemgr_t*, mm_machine_t*);
 void mm_machinemgr_delete(mm_machinemgr_t*, mm_machine_t*);
 mm_machine_t*
-mm_machinemgr_delete_by_id(mm_machinemgr_t*, int);
+mm_machinemgr_delete_by_id(mm_machinemgr_t*, uint64_t);
 
 #endif /* MM_MACHINE_MGR_H */
